separate the keycloak setup in different cucumber feature
