+++ rest-server/src/test/resources/cucumber/location-rest-api.feature
@@ -1,20 +1,9 @@
-Feature: End to End test for REST API authentication
-
-  Background:
-    Given REST server url in system property "rest-server-url"
-    Given Keycloak auth server url in system property "keycloak.url"
-    Given Keycloak admin user "keycloak-admin" with password "admin"
-    Given Keycloak master realm client-id "admin-cli"
-    Then Initial Keycloak utils
-    Then Create Keycloak realm "opennms"
-    Then Add roles
-      | admin |
-      | user  |
-    Then Add admin user "admin-user" with password "password123" and role "admin"
-    Then Add regular user "test-user" with password "password123" and role "user"
+Feature: End to End test for location REST API endpoints with auth
 
   Scenario: admin user can view/add/update/delete location
-    #Given Admin user "admin-user" with password "password123"
+    Given REST server url in system property "rest-server-url"
+    Given Keycloak auth server url in system property "keycloak.url", realm "opennms" and client "admin-cli"
+    Given Admin user "admin-user" with password "password123"
     Then Admin user can create an access token
     Then Admin user can create new location
     Then Admin user can list location
